fix errors when creating new user
