+++ administrator/com_joomgallery/language/en-GB/com_joomgallery.migration.Jg3ToJg4.ini
@@ -15,6 +15,10 @@ FILES_JOOMGALLERY_MIGRATION_JG3TOJG4_THUMBPATH_DESC="Path to thumbnail image"
 FILES_JOOMGALLERY_MIGRATION_IMAGE_TITLE="Migration: Images"
 FILES_JOOMGALLERY_MIGRATION_CATEGORY_TITLE="Migration: Categories"
 FILES_JOOMGALLERY_MIGRATION_CATIMAGE_TITLE="Adjustment: Category thumbnails"
+FILES_JOOMGALLERY_MIGRATION_USER_TITLE="Migration: Users"
+FILES_JOOMGALLERY_MIGRATION_FAVOURITE_TITLE="Migration: Favorites"
+FILES_JOOMGALLERY_MIGRATION_VOTE_TITLE="Migration: Votes"
+FILES_JOOMGALLERY_MIGRATION_COMMENT_TITLE="Migration: Comments"
 FILES_JOOMGALLERY_MIGRATION_CHECK_IMAGE_FILENAMES_TITLE="Image filenames"
 FILES_JOOMGALLERY_MIGRATION_CHECK_IMAGE_FILENAMES_DESC="There are different filenames for thumbnail and detail image defined in the database of your JG3 tables. Number of affected images: %s"
 FILES_JOOMGALLERY_MIGRATION_CHECK_IMAGE_FILENAMES_HELP="It was detected that for some images in the database table of the JG3 (source, #__com_joomgallery) different filenames are stored in the rows 'imgfilename' and 'imgthumbname'. This script could result in errorous migration results. We strongly advise to clean up the database manually before starting the migration. For all the affected Images you have to do the following:<ul><li>Rename the thumbnail image in the filesystem to be the one set in the old JoomGallery image table at row 'imgfilename'.</li><li>In the old JoomGallery image table, set the value of the row 'imgthumbname' to be the same as in row 'imgfilename'.</li></ul><br/>ID's of affected images: (%s)"
