+++ administrator/com_joomgallery/src/Table/FaultyTable.php
@@ -1,7 +1,7 @@
 <?php
 /**
 ******************************************************************************************
-**   @version    4.0.0-dev                                                                  **
+**   @version    4.0.0-dev                                                              **
 **   @package    com_joomgallery                                                        **
 **   @author     JoomGallery::ProjectTeam <team@joomgalleryfriends.net>                 **
 **   @copyright  2008 - 2023  JoomGallery::ProjectTeam                                  **
@@ -13,16 +13,17 @@ namespace Joomgallery\Component\Joomgallery\Administrator\Table;
 // No direct access
 defined('_JEXEC') or die;
 
+use \Joomla\CMS\Factory;
 use \Joomla\CMS\Table\Table;
 use \Joomla\Database\DatabaseDriver;
 
 /**
- * Comments table
+ * Faulty table
  *
  * @package JoomGallery
  * @since   4.0.0
  */
-class CommentsTable extends Table
+class FaultyTable extends Table
 {
   use JoomTableTrait;
 
@@ -33,9 +34,9 @@ class CommentsTable extends Table
 	 */
 	public function __construct(DatabaseDriver $db)
 	{
-		$this->typeAlias = _JOOM_OPTION.'.comments';
+		$this->typeAlias = _JOOM_OPTION.'.faulty';
 
-		parent::__construct(_JOOM_TABLE_COMMENTS, 'id', $db);
+		parent::__construct(_JOOM_TABLE_FAULTIES, 'id', $db);
 	}
 
   /**
@@ -53,41 +54,10 @@ class CommentsTable extends Table
 	public function bind($array, $ignore = '')
 	{
 		$date = Factory::getDate();
-		$task = Factory::getApplication()->input->get('task', '', 'cmd');
-
-    // Support for title field: title
-    if(\array_key_exists('title', $array))
-    {
-      $array['title'] = \trim($array['title']);
-      if(empty($array['title']))
-      {
-        $array['title'] = 'Unknown';
-      }
-    }
 
 		if($array['id'] == 0)
 		{
 			$array['created_time'] = $date->toSql();
-		}
-
-		if(!\key_exists('created_by', $array) || empty($array['created_by']))
-		{
-			$array['created_by'] = Factory::getApplication()->getIdentity()->id;
-		}
-
-		if($task == 'apply' || \strpos($task, 'save') !== false)
-		{
-			$array['modified_time'] = $date->toSql();
-		}
-
-		if($array['id'] == 0 && empty($array['modified_by']))
-		{
-			$array['modified_by'] = Factory::getApplication()->getIdentity()->id;
-		}
-
-		if($task == 'apply' || \strpos($task, 'save') !== false)
-		{
-			$array['modified_by'] = Factory::getApplication()->getIdentity()->id;
 		}
 
 		return parent::bind($array, $ignore);
