--- administrator/com_joomgallery/src/Table/VotesTable.php
@@ -1,69 +1,0 @@
-<?php
-/**
-******************************************************************************************
-**   @version    4.0.0-dev                                                                  **
-**   @package    com_joomgallery                                                        **
-**   @author     JoomGallery::ProjectTeam <team@joomgalleryfriends.net>                 **
-**   @copyright  2008 - 2023  JoomGallery::ProjectTeam                                  **
-**   @license    GNU General Public License version 3 or later                          **
-*****************************************************************************************/
-
-namespace Joomgallery\Component\Joomgallery\Administrator\Table;
-
-// No direct access
-defined('_JEXEC') or die;
-
-use \Joomla\CMS\Table\Table;
-use \Joomla\Database\DatabaseDriver;
-
-/**
- * Votes table
- *
- * @package JoomGallery
- * @since   4.0.0
- */
-class VotesTable extends Table
-{
-  use JoomTableTrait;
-
-	/**
-	 * Constructor
-	 *
-	 * @param   JDatabase  &$db  A database connector object
-	 */
-	public function __construct(DatabaseDriver $db)
-	{
-		$this->typeAlias = _JOOM_OPTION.'.votes';
-
-		parent::__construct(_JOOM_TABLE_VOTES, 'id', $db);
-	}
-
-  /**
-	 * Overloaded bind function to pre-process the params.
-	 *
-	 * @param   array  $array   Named array
-	 * @param   mixed  $ignore  Optional array or list of parameters to ignore
-	 *
-	 * @return  boolean  True on success.
-	 *
-	 * @see     Table:bind
-	 * @since   4.0.0
-	 * @throws  \InvalidArgumentException
-	 */
-	public function bind($array, $ignore = '')
-	{
-		$date = Factory::getDate();
-
-		if($array['id'] == 0)
-		{
-			$array['created_time'] = $date->toSql();
-		}
-
-    if($array['id'] == 0 && (!\key_exists('created_by', $array) || empty($array['created_by'])))
-		{
-			$array['created_by'] = Factory::getApplication()->getIdentity()->id;
-		}
-
-		return parent::bind($array, $ignore);
-	}
-}
